+++ src/agents/auth-profiles/external-oauth.test.ts
@@ -124,7 +124,7 @@ describe("auth external oauth helpers", () => {
     expect(shouldPersist).toBe(true);
   });
 
-  it("overlays external CLI OAuth only when the stored credential is no longer usable", () => {
+  it("does not use Codex CLI OAuth as a runtime overlay source", () => {
     readCodexCliCredentialsCachedMock.mockReturnValue(
       createCredential({
         access: "fresh-cli-access-token",
@@ -146,9 +146,9 @@ describe("auth external oauth helpers", () => {
     );
 
     expect(overlaid.profiles["openai-codex:default"]).toMatchObject({
-      access: "fresh-cli-access-token",
-      refresh: "fresh-cli-refresh-token",
-      expires: expect.any(Number),
+      access: "stale-store-access-token",
+      refresh: "stale-store-refresh-token",
+      accountId: "acct-cli",
     });
   });
